relicense under MIT license
There were several discussions why Leeroy is not licensed under
the terms of the MIT license. According to my knowledge the 2
clause BSD license I usually choose for projects is equivalent to
the MIT license. To clear up confusions and not require everyone
to read all 9 sentences in License.md Leeroy is relicensed under
the terms of the MIT license.